% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/worldcities.R
\docType{data}
\name{worldcities}
\alias{worldcities}
\title{41,000 cities and associated metadata}
\format{
A data frame with 41001 rows and 11 variables:
\describe{
  \item{city}{name of the city, using only ascii characters}
  \item{lat}{latitude coordinates of the city}
  \item{lng}{longitude coordinates of the city}
  \item{country}{country the city is located in, using only ascii characters}
  \item{iso2}{the alpha-2 iso code of the country}
  \item{iso3}{the alpha-3 iso code of the country}
  \item{admin_name}{the name of the highest level administration region of the city town (e.g. a US state or Canadian province). possibly blank. using only ascii characters}
  \item{capital}{blank string if not a capital, otherwise: primary - countrys capital (e.g. Washington D.C.), admin - first-level admin capital (e.g. Little Rock, AR), minor - lower-level admin capital (e.g. Fayetteville, AR)}
  \item{population}{cities urban population. if city urban population is not available, the municipal population is used}
  \item{id}{A 10-digit unique id generated by SimpleMaps}
}
}
\source{
\url{https://simplemaps.com/data/world-cities}
}
\usage{
worldcities
}
\description{
A dataset containing coordinates, populations, and country for 41,000 worldwide cities
}
\details{
All data comes from the World Cities dataset contained at SimpleMaps (https://simplemaps.com/data/world-cities)
}
\keyword{datasets}
